Split features & steps into distinct files
