+++ source/core/tests/integration/features/stream_submitted_silver_measurements_to_gold_measurements.feature
@@ -1,0 +1,13 @@
+Feature: Streaming Submitted Silver Measurements to Gold Measurements
+
+  Scenario: Processing Silver measurements to Gold
+    Given valid measurements inserted into the silver measurements table
+    When streaming Silver measurements to the Gold layer
+    Then measurements are available in the gold measurements table
+
+  Scenario: Creating receipt entry during Silver to Gold streaming
+    Given valid measurements with an orchestration instance id inserted into the silver measurements table
+    When streaming Silver measurements to the Gold layer
+    Then a receipt entry is available in the process manager receipts table
+
+
